+++ .stylelintrc.js
@@ -6,11 +6,13 @@ module.exports = {
   plugins: [
     'stylelint-css-modules',
     './src/linters/colors.js',
+    './src/linters/typography.js',
   ],
   rules: {
     'css-modules/composed-class-names': true,
     'css-modules/css-variables': true,
     'no-descending-specificity': null,
     'mds/colors': true,
+    'mds/typography': true,
   },
 };